feat(briefing): extract shared/ utilities (staleness, market lookup, JSON parser)
Workstream 6 Step 1, Commit 2/11 — pure relocation of three reusable
helpers into a shared/ subdirectory so future pipeline modules (and the
future aggregator) import from a single source of truth instead of
each grabbing internals from briefing-service.js.

Created:
- server/lib/briefing/shared/safe-json-parse.js (~210 lines)
  Multi-attempt LLM JSON parser hardened against markdown fences,
  citations, single-quote delimiters, unquoted keys, embedded // comments,
  malformed brackets. Pure function, zero deps. Used by 11 call sites.
- server/lib/briefing/shared/staleness.js (~85 lines)
  isDailyBriefingStale, isEventsStale, isTrafficStale, areEventsEmpty —
  pure cache-staleness predicates used by getOrGenerateBriefing.
- server/lib/briefing/shared/get-market-for-location.js (~50 lines)
  City/state → rideshare market name (e.g., Plano,TX → Dallas).
  Used by ≥2 pipelines (events, news).

In briefing-service.js:
- Removed 6 local definitions (~290 lines)
- Added 3 imports
- Replaced each removed function with a one-line relocation comment

Behavior parity: identical. All call sites resolve to the same code via
import; no signatures or semantics changed.

Per docs/review-queue/PLAN_workstream6_briefing_split.md §3.5.
